refactor(ui): Move delete_item to common/delete_item
