+++ apps/clearance_ui/src/components/clearance_library/license_conclusions/ActionCell.tsx
@@ -17,8 +17,8 @@ import {
     TooltipTrigger,
 } from "@/components/ui/tooltip";
 import { toast } from "@/components/ui/use-toast";
+import DeleteLicenseConclusion from "@/components/common/delete_item/DeleteLicenseConclusion";
 import EditButton from "@/components/common/edit_item/EditButton";
-import DeleteLicenseConclusion from "@/components/delete_item/DeleteLicenseConclusion";
 import { isValidConcludedExpression } from "@/helpers/isValidConcludedExpression";
 
 type LicenseConclusion = ZodiosResponseByAlias<
