+++ apps/clearance_ui/src/components/common/delete_item/DeleteDialog.tsx
